mudanças nas classes designadas no notion

+++ src/main/java/com/vertex/vertex/personalization/service/PersonalizationService.java
@@ -8,28 +8,42 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 @AllArgsConstructor
 public class PersonalizationService {
+
     private final PersonalizationRepository personalizationRepository;
 
-    public Personalization save(Personalization group){
-        return personalizationRepository.save(group);
+    public Personalization save(Personalization personalization) {
+        return personalizationRepository.save(personalization);
     }
 
-    public List<Personalization> findAll(){
+    public Personalization put(Personalization personalization) {
+        if (personalization.getId()!= null && !personalizationRepository.existsById(personalization.getId())) {
+            throw new RuntimeException("A personalização não existe.");
+        }
+        return personalizationRepository.save(personalization);
+    }
+
+    public List<Personalization> findAll() {
         return personalizationRepository.findAll();
     }
 
-    public Personalization findById(Long id){
+    public Personalization findById(Long id) {
+        if (!personalizationRepository.existsById(id)) {
+            throw new NoSuchElementException("Personalização escolhida não existe.");
+        }
         return personalizationRepository.findById(id).get();
     }
 
-    public void deleteById(Long id){
-        personalizationRepository.deleteById(id);
+    public void deleteById(Long id) {
+        if (!personalizationRepository.existsById(id)) {
+            throw new NoSuchElementException("Personalização escolhida para deletar não existe.");
+        } else {
+            personalizationRepository.deleteById(id);
+        }
     }
-
-
 
 }
